+++ rust/CHANGELOG.md
@@ -4,3 +4,8 @@
 
 - Fixed the --secondary-fs argument
   ([#181](https://github.com/saidsay-so/pjdfstest/pull/181))
+
+### Changed
+
+- Removed the need for a third user from the config file.
+  ([#180](https://github.com/saidsay-so/pjdfstest/pull/180))